+++ src/overrides/prettier.js
@@ -1,9 +1,12 @@
+import tailwindcss from 'prettier-plugin-tailwindcss'
+
 /** @type { import('prettier').Config } */
 const prettierConfig = {
   printWidth: 120,
   semi: false,
   singleQuote: true,
   jsxSingleQuote: true,
+  plugins: [tailwindcss],
 }
 
 /** @type { import('eslint').Linter.RulesRecord } */