+++ changelog.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.6.0] - 2023-04-26
 ### Added
 - Add support for Next.js
 - Add support for Tailwind CSS
@@ -60,7 +62,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - First release!
 
-[unreleased]: https://github.com/arianrhodsandlot/eslint-config/compare/v0.5.0...HEAD
+[unreleased]: https://github.com/arianrhodsandlot/eslint-config/compare/v0.6.0...HEAD
+[0.6.0]: https://github.com/arianrhodsandlot/eslint-config/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/arianrhodsandlot/eslint-config/compare/v0.4.2...v0.5.0
 [0.4.2]: https://github.com/arianrhodsandlot/eslint-config/compare/v0.4.1...v0.4.2
 [0.4.1]: https://github.com/arianrhodsandlot/eslint-config/compare/v0.3.1...v0.4.1